requirements.txt
flask
flask-cors
python-dotenv
supabase
postgrest
httpx
storage3
realtime
